backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pdfplumber==0.10.3
python-docx==1.1.0
joblib==1.3.2
shap==0.44.0
lime==0.2.0.1